Move ForceList to its own file
